Fix type mismatches and missing prototypes in gui
Address compiler warnings by fixing type mismatches and adding missing
function prototypes:

- Add missing prototype for display() function in gui_private.h
- Change vk_special_time from int to unsigned int to match 'now' type
  (both are timestamps, fixes sign conversion warnings)
- Change 'start' variable from long long to uint64_t to match
  SDL_GetTicks64() return type
- Change 't' variable from int to time_t to match time() return type
- Change 'duration' from int to int64_t for time difference calculations
- Fix size calculations in display_vc code: perform calculations in
  wider types (uint64_t) and clamp to bargraph range (42) before
  casting to int, avoiding information loss
- Fix format specifier for time_t (%ld instead of %d)

All changes preserve the intended behavior while eliminating implicit
conversions that could lose information.

+++ src/gui/gui_input.c
@@ -462,7 +462,7 @@ void gui_sdl_mouseproc(int x, int y, int what)
 				vk_special_dec();
 				delta++;
 			}
-			vk_special_time = (int)now;
+			vk_special_time = now;
 
 			if (mdown) {
 				shift_override = special_tab[vk_special].shift_over;
